app: split each command executor into its own source
